Fix PR review findings: cmap offset overflow, glyf flag-repeat clip, cmap glyph clamp
- CmapTable.cs: use overflow-safe (long) arithmetic for the subtable-offset
  bounds check in Parse (an offset near int.MaxValue previously wrapped
  around and passed the check, leading to an IndexOutOfRangeException
  downstream), and apply the same overflow-safe pattern to the equivalent
  entry-point bounds checks in TryParseFormat4/TryParseFormat12.
- GlyfLocaReader.cs: reject a simple-glyph flag repeat count that exceeds
  the remaining point slots with InvalidDataException instead of silently
  clipping it to the remaining count, matching this class's documented
  malformed-input contract.
- TrueTypeFont.cs: clamp GetGlyphIndex's result to 0 when the cmap subtable
  maps a codepoint to a glyph index outside [0, GlyphCount), so callers can
  always safely pass its result straight to GetGlyphOutline; updated the
  XML doc comment to make this guarantee explicit.
- Added/extended unit tests reproducing each malformed-input scenario using
  SyntheticFontBuilder synthetic fonts (no third-party font fixtures).

+++ src/DemaConsulting.CanvasNet/Fonts/CmapTable.cs
@@ -70,7 +70,7 @@ internal sealed class CmapTable
             var encodingId = SfntContainer.ReadUInt16(data, recordOffset + 2);
             var subtableOffset = SfntContainer.ReadInt32(data, recordOffset + 4);
 
-            if (subtableOffset < 0 || subtableOffset + 2 > tableLength)
+            if (subtableOffset < 0 || (long)subtableOffset + 2 > tableLength)
             {
                 continue;
             }
@@ -152,7 +152,7 @@ internal sealed class CmapTable
     /// </summary>
     private static Func<int, int>? TryParseFormat4(byte[] data, int offset, int tableEnd)
     {
-        if (offset + 14 > tableEnd)
+        if ((long)offset + 14 > tableEnd)
         {
             return null;
         }
@@ -231,7 +231,7 @@ internal sealed class CmapTable
     /// </summary>
     private static Func<int, int>? TryParseFormat12(byte[] data, int offset, int tableEnd)
     {
-        if (offset + 16 > tableEnd)
+        if ((long)offset + 16 > tableEnd)
         {
             return null;
         }
